feat(init, core, treesitter, ui, user, utils): Update lua configs and add/remove  plugins
Paths and dependencies in `init.lua` and `core/init.lua` are updated and
core.autocmd and core.autocmd_optional are enabled.
Plugin `HiPhish/nvim-ts-rainbow2` is added in `treesitter/plugins.lua` with
 its
respective configuration, specially highlighting the setup of rainbow
parens and the disabling of the plugin for jsx and cpp languages.

In UI related configs, `heirline.lua` has an additional buffer match configuration
,while plugin `mawkler/modicator.nvim` is removed in `ui/plugins.lua` and plugins
 `mvllow/modes.nvim`,
`tzachar/highlight-undo.nvim`, and `3rd/image.nvim` are updated or added with
 their corresponding configs.

Lastly, the 'utils/ui/highlights.lua' has a small update on 'CursorLineNr' attribute