refactor(lsp, latex, lsp-plugins, documentation): Update configurations, improve  latex utilities and remove unused plugins
This commit refactors several areas of the project for better functionality
.
The 'lsp' configuration in the 'config.lua' file was updated, the ErrorLens
Telescope was moved from 'LspAttach' to 'VeryLazy', and 'texlab' was condition
ally implemented for LaTeX.

The 'latex' modules were also modified. The vimtex view options were updated
 and multiple new functions were added in the 'config.lua' file. In the 'plugins
.lua' file, the 'lazy' attribute was removed from multiple plugins, and 'lt
ex-extra.nvim' was conditionally implemented based on 'latex' setting.

In 'documentation', the monaqa/nvim-treesitter-clipping plugin was removed from
 the 'plugins.lua' file.

The list of LSP-plugins was updated in the 'plugins.lua' file to remove the
 unused 'right-corner-diagnostics.nvim' and add 'diagflow.nvim'. The setups
 for LspLensOn, LspLensOff, and LspLensToggle were also adjusted.

The 'jump.lua' file was updated to initialize patterns for the toggle flash
 search keybind with any charachter and the keybind was changed from '<leader
>/' to '\\/'.